+++ explorer-api/src/mix_nodes/mod.rs
@@ -6,7 +6,6 @@ use std::time::Duration;
 pub(crate) mod http;
 pub(crate) mod location;
 pub(crate) mod models;
-pub(crate) mod utils;
 
 pub(crate) const CACHE_REFRESH_RATE: Duration = Duration::from_secs(30);
 pub(crate) const CACHE_ENTRY_TTL: Duration = Duration::from_secs(60);