+++ nym-wallet/src/components/Modals/SimpleModal.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Box, Button, Modal, Stack, SxProps, Typography } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
 import ErrorOutline from '@mui/icons-material/ErrorOutline';
+import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
 import { StyledBackButton } from 'src/components/StyledBackButton';
 import { modalStyle } from './styles';
 
@@ -9,8 +10,10 @@ export const SimpleModal: React.FC<{
   open: boolean;
   hideCloseIcon?: boolean;
   displayErrorIcon?: boolean;
+  displayInfoIcon?: boolean;
   headerStyles?: SxProps;
   subHeaderStyles?: SxProps;
+  buttonFullWidth?: boolean;
   onClose?: () => void;
   onOk?: () => Promise<void>;
   onBack?: () => void;
@@ -24,8 +27,10 @@ export const SimpleModal: React.FC<{
   open,
   hideCloseIcon,
   displayErrorIcon,
+  displayInfoIcon,
   headerStyles,
   subHeaderStyles,
+  buttonFullWidth,
   onClose,
   okDisabled,
   onOk,
@@ -38,8 +43,9 @@ export const SimpleModal: React.FC<{
   backdropProps,
 }) => (
   <Modal open={open} onClose={onClose} BackdropProps={backdropProps}>
-    <Box sx={{ ...modalStyle, ...sx }}>
+    <Box sx={{ border: (t) => `1px solid ${t.palette.nym.nymWallet.modal.border}`, ...modalStyle, ...sx }}>
       {displayErrorIcon && <ErrorOutline color="error" sx={{ mb: 3 }} />}
+      {displayInfoIcon && <InfoOutlinedIcon sx={{ mb: 2, color: (theme) => theme.palette.nym.nymWallet.text.blue }} />}
       <Stack direction="row" justifyContent="space-between" alignItems="center">
         {typeof header === 'string' ? (
           <Typography fontSize={20} fontWeight={600} sx={{ color: 'text.primary', ...headerStyles }}>
@@ -52,7 +58,7 @@ export const SimpleModal: React.FC<{
       </Stack>
 
       <Typography
-        mt={0.5}
+        mt={subHeader ? 0.5 : 0}
         mb={3}
         fontSize={12}
         color={(theme) => theme.palette.text.secondary}
@@ -64,10 +70,10 @@ export const SimpleModal: React.FC<{
       {children}
 
       {(onOk || onBack) && (
-        <Box sx={{ display: 'flex', alignItems: 'center', gap: 3, mt: 3 }}>
+        <Box sx={{ display: 'flex', alignItems: 'center', gap: 2, mt: 2, width: buttonFullWidth ? '100%' : null }}>
           {onBack && <StyledBackButton onBack={onBack} />}
           {onOk && (
-            <Button variant="contained" fullWidth size="large" onClick={onOk} disabled={okDisabled}>
+            <Button variant="contained" fullWidth size="large" onClick={onOk} disabled={okDisabled} sx={{ mt: 3 }}>
               {okLabel}
             </Button>
           )}