+++ ts-packages/types/src/types/rust/UnbondedMixnode.ts
@@ -2,5 +2,5 @@ export interface UnbondedMixnode {
   identity_key: string;
   owner: string;
   proxy: string | null;
-  unbonding_height: bigint;
+  unbonding_height: number;
 }